moved EquipmentCard to folder container
